+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.3
+
+- Add extension for decode strings to json.
+- Add constant for empty Json.
+
 ## 0.0.2
 
 - Fix examples in readme and example file.